+++ bin-src/types.ts
@@ -129,6 +129,7 @@ export interface Context {
   skipSnapshots?: boolean;
   now?: number;
   startedAt?: number;
+  activity?: { end: () => void };
   exitCode: number;
   exitCodeKey: string;
   userError?: boolean;
@@ -140,6 +141,7 @@ export interface Context {
   closeTunnel?: () => void;
   isPublishOnly?: boolean;
   isOnboarding: boolean;
+  turboSnapAvailability?: string;
 
   http: {
     fetch: (url: string, options?: RequestInit, opts?: any) => Promise<Response>;
@@ -200,6 +202,10 @@ export interface Context {
     status: string;
     autoAcceptChanges: boolean;
     reportToken: string;
+    app: {
+      id: string;
+      turboSnapAvailability: string;
+    };
   };
   build: {
     id: string;
@@ -219,9 +225,11 @@ export interface Context {
     interactionTestFailuresCount: number;
     inProgressCount?: number;
     autoAcceptChanges: boolean;
+    turboSnapEnabled?: boolean;
     wasLimited?: boolean;
     startedAt?: number;
     app: {
+      manageUrl: string;
       setupUrl: string;
       account?: {
         exceededThreshold: boolean;
@@ -261,6 +269,7 @@ export interface Context {
   };
   uploadedBytes?: number;
   turboSnap?: Partial<{
+    unavailable?: boolean;
     rootPath: string;
     baseDir: string;
     storybookDir: string;