+++ bin-src/ui/messages/warnings/turboSnapUnavailable.stories.ts
@@ -1,0 +1,14 @@
+import turboSnapUnavailable from './turboSnapUnavailable';
+
+export default {
+  title: 'CLI/Messages/Warnings',
+};
+
+export const TurboSnapUnavailable = () =>
+  turboSnapUnavailable({
+    build: {
+      app: {
+        manageUrl: 'https://www.chromatic.com/manage?appId=59c59bd0183bd100364e1d57',
+      },
+    },
+  });